a70q: Pass call state to audio hal and patch RIL
Simon1511: libsec-ril sets a prop (ril.dds.call.slotid) when in call with Sim1 or Sim2. Patch libsec-ril to make that prop vendor lable-able (vendor.calls.slotid). Then use that prop to check which Sim is currently in a call (0 for Sim1 and 1 for Sim2) and pass the arguments to audio HAL depending on that.

Note for later: Check whether passing [g_]call_state is really needed

Change-Id: Ibb0f25432cd3692f73e4196ec6ab0d07bbde4640